Move getWheelhouses functions to utils.js. Export functions and import them where they are used.
